Swapped over to using decimals for layout. This fixes annoying layout errors caused by floating point imprecision

+++ BeautifulBlueprints/Elements/Space.cs
@@ -8,12 +8,12 @@ namespace BeautifulBlueprints.Elements
     {
         public Space(
             string name = null,
-            float minWidth = DEFAULT_MIN_WIDTH,
-            float? preferredWidth = null,
-            float maxWidth = DEFAULT_MAX_WIDTH,
-            float minHeight = DEFAULT_MIN_HEIGHT,
-            float? preferredHeight = null,
-            float maxHeight = DEFAULT_MAX_HEIGHT
+            decimal minWidth = DEFAULT_MIN_WIDTH,
+            decimal? preferredWidth = null,
+            decimal maxWidth = DEFAULT_MAX_WIDTH,
+            decimal minHeight = DEFAULT_MIN_HEIGHT,
+            decimal? preferredHeight = null,
+            decimal maxHeight = DEFAULT_MAX_HEIGHT
         )
             : base(name, minWidth, preferredWidth, maxWidth, minHeight, preferredHeight, maxHeight)
         {
@@ -27,7 +27,7 @@ namespace BeautifulBlueprints.Elements
         /// <param name="top"></param>
         /// <param name="bottom"></param>
         /// <returns></returns>
-        internal override IEnumerable<Solver.Solution> Solve(float left, float right, float top, float bottom)
+        internal override IEnumerable<Solver.Solution> Solve(decimal left, decimal right, decimal top, decimal bottom)
         {
             yield return FillSpace(left, right, top, bottom);
         }
